requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
urllib3==2.2.3
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.51.0
pydantic==2.9.2
tenacity==9.0.0
schedule==1.2.2
orjson==3.10.7
numpy==2.1.3
httpx==0.27.2
fastembed==0.7.3
transformers==4.36.0
torch==2.1.0
